Adjust DO tags
