--- src/main/java/com/ldtteam/domumornamentum/datagen/global/ConcreteTagProvider.java
@@ -1,51 +1,0 @@
-package com.ldtteam.domumornamentum.datagen.global;
-
-import com.ldtteam.domumornamentum.tag.ModTags;
-import com.ldtteam.domumornamentum.util.Constants;
-import net.minecraft.data.DataGenerator;
-import net.minecraft.data.tags.BlockTagsProvider;
-import net.minecraft.world.level.block.Blocks;
-import net.minecraftforge.common.data.ExistingFileHelper;
-import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
-
-public class ConcreteTagProvider extends BlockTagsProvider
-{
-    public ConcreteTagProvider(
-      final DataGenerator generatorIn,
-      @Nullable final ExistingFileHelper existingFileHelper)
-    {
-        super(generatorIn, Constants.MOD_ID, existingFileHelper);
-    }
-
-    @SuppressWarnings("unchecked")
-    @Override
-    protected void addTags()
-    {
-        this.tag(ModTags.CONCRETE)
-          .add(
-            Blocks.BLACK_CONCRETE,
-            Blocks.CYAN_CONCRETE,
-            Blocks.BLUE_CONCRETE,
-            Blocks.BROWN_CONCRETE,
-            Blocks.GRAY_CONCRETE,
-            Blocks.GREEN_CONCRETE,
-            Blocks.LIGHT_BLUE_CONCRETE,
-            Blocks.LIGHT_GRAY_CONCRETE,
-            Blocks.LIME_CONCRETE,
-            Blocks.MAGENTA_CONCRETE,
-            Blocks.ORANGE_CONCRETE,
-            Blocks.PINK_CONCRETE,
-            Blocks.PURPLE_CONCRETE,
-            Blocks.RED_CONCRETE,
-            Blocks.WHITE_CONCRETE,
-            Blocks.YELLOW_CONCRETE);
-    }
-
-    @Override
-    @NotNull
-    public String getName()
-    {
-        return "Concrete Tag Provider";
-    }
-}
