+++ src/main/java/com/ldtteam/domumornamentum/datagen/slab/SlabComponentTagProvider.java
@@ -27,51 +27,38 @@ public class SlabComponentTagProvider extends BlockTagsProvider
 
         this.tag(ModTags.SLAB_MATERIALS)
           .add(
-            Blocks.CLAY,
-            Blocks.TERRACOTTA,
-            Blocks.WHITE_TERRACOTTA,
-            Blocks.ORANGE_TERRACOTTA,
-            Blocks.MAGENTA_TERRACOTTA,
-            Blocks.LIGHT_BLUE_TERRACOTTA,
-            Blocks.YELLOW_TERRACOTTA,
-            Blocks.LIME_TERRACOTTA,
-            Blocks.PINK_TERRACOTTA,
-            Blocks.GRAY_TERRACOTTA,
-            Blocks.LIGHT_GRAY_TERRACOTTA,
-            Blocks.CYAN_TERRACOTTA,
-            Blocks.PURPLE_TERRACOTTA,
-            Blocks.BLUE_TERRACOTTA,
-            Blocks.BROWN_TERRACOTTA,
-            Blocks.GREEN_TERRACOTTA,
-            Blocks.RED_TERRACOTTA,
-            Blocks.BLACK_TERRACOTTA,
-            Blocks.BLACK_TERRACOTTA,
             Blocks.BLACKSTONE,
             Blocks.GILDED_BLACKSTONE,
+            Blocks.NETHERRACK,
+            Blocks.CRIMSON_NYLIUM,
+            Blocks.WARPED_NYLIUM,
+            Blocks.BASALT,
+            Blocks.POLISHED_BASALT,
+            Blocks.SMOOTH_BASALT,
             Blocks.HAY_BLOCK,
             Blocks.COPPER_BLOCK,
             Blocks.CUT_COPPER,
             Blocks.EXPOSED_COPPER,
             Blocks.OXIDIZED_COPPER,
             Blocks.WEATHERED_COPPER,
-            Blocks.WEATHERED_CUT_COPPER,
             Blocks.WAXED_COPPER_BLOCK,
-            Blocks.WAXED_CUT_COPPER,
             Blocks.WAXED_EXPOSED_COPPER,
             Blocks.WAXED_OXIDIZED_COPPER,
-            Blocks.WAXED_WEATHERED_COPPER,
-            Blocks.WAXED_WEATHERED_CUT_COPPER
+            Blocks.WAXED_WEATHERED_COPPER
           )
           .addTags(
-            BlockTags.LOGS,
-            BlockTags.CORAL_BLOCKS,
-            BlockTags.WART_BLOCKS,
             ModTags.EXTRA_BLOCKS,
             ModTags.BRICKS,
+            ModTags.CONCRETE,
+            ModTags.TERRACOTTA,
             BlockTags.DIRT,
             BlockTags.WOOL,
-            ModTags.CONCRETE,
-            Tags.Blocks.GLASS
+            BlockTags.LEAVES,
+            Tags.Blocks.STORAGE_BLOCKS,
+            Tags.Blocks.GLASS,
+            BlockTags.LOGS,
+            BlockTags.CORAL_BLOCKS,
+            BlockTags.WART_BLOCKS
           );
     }
